Move Product and ProductList to pages directory
